bitcoin_witness: Move tests to blockchain_contracts
Without adding dependencies to blockchain_contracts!
