+++ cnd/Cargo.toml
@@ -58,7 +58,6 @@ void = "1.0.2"
 warp = { version = "0.1", default-features = false }
 
 [dev-dependencies]
-bitcoin_rpc_test_helpers = { path = "../internal/bitcoin_rpc_test_helpers" }
 bitcoincore-rpc = "0.8.0-rc1"
 maplit = "1"
 matches = "0.1.8"
