--- internal/bitcoin_rpc_test_helpers/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "bitcoin_rpc_test_helpers"
-version = "0.1.0"
-authors = [ "CoBloX developers <team@coblox.tech>" ]
-edition = "2018"
-
-[dependencies]
-bitcoin = "0.19.1"
-bitcoincore-rpc = "0.8.0-rc1"
-testcontainers = "0.8"
